+++ src/store/types.ts
@@ -1,0 +1,4 @@
+export type CryptoState {
+    cryptos: Crypto[];
+    loading: boolean;
+    }
